syncoff, unit instructions, and a sync fix
The sync fix is most interesting:

Before, if the instr was sync:true, we would insert a sync right
after it, on the same device.  But we computed the contents of the
sync instruction before updating the queue times from the main
instruction, which meant they were all wrong.  Fun!